Refactor section menu
