+++ components/Context.js
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useReducer, useState} from 'react';
+import React, { createContext, useContext, useState} from 'react';
 
 export const AppContext = createContext()
 export const useAppContext = () => useContext(AppContext)
